refactor _to_time_str
