+++ src/imars3d/backend/dataio/data.py
@@ -1,18 +1,24 @@
 #!/usr/bin/env python3
 """Data handling for iMars3D."""
-import re
-from datetime import datetime
-import logging
-import param
-from imars3d.backend.util.functions import clamp_max_workers
-import numpy as np
-import tifffile
-from functools import partial
-from pathlib import Path
-from fnmatch import fnmatchcase
-from typing import Optional, Tuple, List, Callable
-from tqdm.contrib.concurrent import process_map
+
+# package imports
 from imars3d.backend.dataio.metadata import MetaData
+from imars3d.backend.util.functions import clamp_max_workers, to_time_str
+
+# third party imports
+import numpy as np
+import param
+import tifffile
+from tqdm.contrib.concurrent import process_map
+
+# standard imports
+from datetime import datetime
+from functools import partial
+from fnmatch import fnmatchcase
+import logging
+from pathlib import Path
+import re
+from typing import Optional, Tuple, List, Callable
 
 # ignore warnings generated by importing dxchange
 import warnings
@@ -475,22 +481,6 @@ def _extract_rotation_angles(
     return rotation_angles
 
 
-def _to_time_str(value: datetime) -> str:
-    """
-    Convert the supplied datetime to a formatted string.
-
-    Parameters
-    ----------
-    value:
-        datetime object to format correctly
-
-    Returns
-    -------
-        The datetime as YYYYMMDDhhmm
-    """
-    return value.strftime("%Y%m%d%H%M")
-
-
 def _save_data(filename: Path, data: np.ndarray, rot_angles: np.ndarray = None) -> None:
     if data is None:
         raise ValueError("Failed to supply data")
@@ -549,7 +539,7 @@ class save_data(param.ParameterizedFunction):
         if params.data is None:
             raise ValueError("Did not supply data")
 
-        save_dir = Path(params.outputbase) / f"{params.name}_{_to_time_str(datetime.now())}"
+        save_dir = Path(params.outputbase) / f"{params.name}_{to_time_str(datetime.now())}"
 
         # save the data as tiffs
         _save_data(filename=save_dir / params.name, data=params.data, rot_angles=params.rot_angles)
@@ -597,7 +587,7 @@ class save_checkpoint(param.ParameterizedFunction):
         # sanitize arguments
         params = param.ParamOverrides(self, params)
 
-        save_dir = params.outputbase / f"{params.name}_chkpt_{_to_time_str(datetime.now())}"
+        save_dir = params.outputbase / f"{params.name}_chkpt_{to_time_str(datetime.now())}"
 
         # save the data as tiffs
         _save_data(filename=save_dir / params.name, data=params.data, rot_angles=params.rot_angles)
