+++ src/imars3d/backend/util/functions.py
@@ -1,8 +1,10 @@
 #!/usr/bin/env python3
 """Util for imars3d."""
 
-import resource
+# standard imports
+from datetime import datetime
 import multiprocessing
+import resource
 from typing import Union
 
 
@@ -14,3 +16,19 @@ def clamp_max_workers(max_workers: Union[int, None]) -> int:
     if max_workers is None:
         max_workers = 0
     return min(resource.RLIMIT_NPROC, max(1, multiprocessing.cpu_count() - 2)) if max_workers <= 0 else max_workers
+
+
+def to_time_str(value: datetime) -> str:
+    """
+    Convert the supplied datetime to a formatted string.
+
+    Parameters
+    ----------
+    value:
+        datetime object to format correctly
+
+    Returns
+    -------
+        The datetime as YYYYMMDDhhmm
+    """
+    return value.strftime("%Y%m%d%H%M")
